Replace use of the old SnapshotService with the new service
