+++ snapshots/snapshot_generator/src/test/scala/weco/catalogue/snapshot_generator/SnapshotGeneratorFeatureTest.scala
@@ -1,22 +1,18 @@
-package uk.ac.wellcome.platform.snapshot_generator
-
-import java.time.Instant
+package weco.catalogue.snapshot_generator
 
 import com.sksamuel.elastic4s.Index
 import org.scalatest.concurrent.{Eventually, IntegrationPatience}
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
-
 import uk.ac.wellcome.akka.fixtures.Akka
 import uk.ac.wellcome.display.models.{ApiVersions, DisplaySerialisationTestBase}
 import uk.ac.wellcome.fixtures.TestWith
-import uk.ac.wellcome.models.Implicits._
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.json.utils.JsonAssertions
 import uk.ac.wellcome.messaging.fixtures.SQS.Queue
 import uk.ac.wellcome.messaging.memory.MemoryMessageSender
+import uk.ac.wellcome.models.Implicits._
 import uk.ac.wellcome.models.work.generators.WorkGenerators
-import uk.ac.wellcome.platform.snapshot_generator.fixtures.WorkerServiceFixture
 import uk.ac.wellcome.platform.snapshot_generator.models.{
   CompletedSnapshotJob,
   SnapshotJob
@@ -24,6 +20,9 @@ import uk.ac.wellcome.platform.snapshot_generator.models.{
 import uk.ac.wellcome.platform.snapshot_generator.test.utils.S3GzipUtils
 import uk.ac.wellcome.storage.fixtures.S3Fixtures.Bucket
 import uk.ac.wellcome.storage.s3.S3ObjectLocation
+import weco.catalogue.snapshot_generator.fixtures.WorkerServiceFixture
+
+import java.time.Instant
 
 class SnapshotGeneratorFeatureTest
     extends AnyFunSpec
