+++ snapshots/snapshot_generator/src/test/scala/weco/catalogue/snapshot_generator/fixtures/NewSnapshotServiceFixture.scala
@@ -5,14 +5,14 @@ import org.scalatest.Suite
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.models.index.IndexFixtures
 import uk.ac.wellcome.storage.fixtures.S3Fixtures
-import weco.catalogue.snapshot_generator.NewSnapshotService
+import weco.catalogue.snapshot_generator.services.SnapshotService
 
 trait NewSnapshotServiceFixture extends IndexFixtures with S3Fixtures { this: Suite =>
   def withSnapshotService[R](worksIndex: Index = "worksIndex",
                              elasticClient: ElasticClient = elasticClient)(
-                              testWith: TestWith[NewSnapshotService, R]): R =
+                              testWith: TestWith[SnapshotService, R]): R =
     testWith(
-      new NewSnapshotService(
+      new SnapshotService(
         index = worksIndex,
       )(elasticClient, s3Client)
     )
